Rename itemWhitelist -> allowedItems
