+++ server/src/allowedItems.ts
